Add Comment feature, and all DB Tables

+++ Models/Achievement.cs
@@ -2,7 +2,7 @@
 
 public partial class Achievement
 {
-    public int Id { get; set; }
+    public ushort Id { get; set; }
 
     public string Name { get; set; } = null!;
